랭킹 기능 변경, 좌표 관련 전부 float -> double로 수정

+++ back-end/plogging/src/main/java/com/backend/plogging/domain/Image.java
@@ -31,9 +31,9 @@ public class Image {
     private LocalDateTime createdAt;
 
     @Column(name = "img_lat")
-    private Float imgLat;
+    private Double imgLat;
 
     @Column(name = "img_lng")
-    private Float imgLng;
+    private Double imgLng;
 
 }
